requirements.txt
Flask
Flask-RESTful
python-dotenv
Flask-JWT-Extended
passlib
bcrypt
psycopg2-binary
sqlalchemy
unidecode
flask-swagger-ui
